Allow shallow clones to fetch more commits on `resetHard` if needed
The behaviour mimics that of the core GoCD `GitMaterial`. By default shallow clones
have 2 commits. If the reset-to revision cannot be found; an additional 98 commits are
fetched. If it is still not found; the entire repo is fetched (effectively unshallowing the clone).
Change URL validation to not require `.git` suffix

These are not mandatory in the real world; especially when HTTP-hosted. For example AWS
CodeCommit URLs do not canonically include `.git` in them. I note that
other such GoCD plugins basically disable validation and rely on connection
testing instead. Perhaps we should do this here.